POC that uses a mysql client to fetch data
This doesn't have any support for credentials and connection info.
This client (and most I've seen) don't support things like
mysql_config_editor or .my.cnf files, which limits the functionality of
the mysqladmin
Additionally this client is only useful for a live server, the old
parser still has to exist for static files.
This format is simpler and seems more performant than raw parsing, but
I don't have the benchmark tests to prove it.